+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 # Head
 
+-   Changed: rename options from `template` to `templates`.
 -   Refactor: simplify code.
 
 # 3.0.0 - 2017-06-20